requirements.txt
# Core Data Science
numpy==1.26.4
pandas==2.2.1
scipy==1.13.0

# Machine Learning
scikit-learn==1.4.2
xgboost==2.0.3

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.20.0

# Jupyter
jupyter==1.0.0
notebook==7.1.3
ipykernel==6.29.4

# EDA Profiling
ydata-profiling==4.7.0

# Data Download
kaggle==1.6.6
openpyxl==3.1.2

# Model Persistence
joblib==1.4.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
